# Reduce runtime overhead when extension is disabled