--- a/src/base/base.c
+++ b/src/base/base.c
@@ -666,11 +666,6 @@ function_stack_entry *xdebug_add_stack_frame(zend_execute_data *zdata, zend_op_a
 
 static void xdebug_execute_ex(zend_execute_data *execute_data)
 {
-	if (EXPECTED(XG(globals.library.mode) == XDEBUG_MODE_OFF)) {
-		xdebug_old_execute_ex(execute_data);
-		return;
-	}
-
 	zend_op_array        *op_array = &(execute_data->func->op_array);
 	zend_execute_data    *edata = execute_data->prev_execute_data;
 	function_stack_entry *fse;
@@ -678,6 +673,11 @@ static void xdebug_execute_ex(zend_execute_data *execute_data)
 	char                 *code_coverage_function_name = NULL;
 	zend_string          *code_coverage_filename = NULL;
 	int                   code_coverage_init = 0;
+
+	if (EXPECTED(XG(globals.library.mode) == XDEBUG_MODE_OFF)) {
+		xdebug_old_execute_ex(execute_data);
+		return;
+	}
 
 	/* For PHP 7, we need to reset the opline to the start, so that all opcode
 	 * handlers are being hit. But not for generators, as that would make an
@@ -846,15 +846,6 @@ static int check_soap_call(function_stack_entry *fse, zend_execute_data *execute
 
 static void xdebug_execute_internal(zend_execute_data *current_execute_data, zval *return_value)
 {
-	if (XG(globals.library.mode) == XDEBUG_MODE_OFF) {
-		if (xdebug_old_execute_internal) {
-			xdebug_old_execute_internal(current_execute_data, return_value);
-		} else {
-			execute_internal(current_execute_data, return_value);
-		}
-		return;
-	}
-
 	zend_execute_data    *edata = EG(current_execute_data);
 	function_stack_entry *fse;
 	int                   function_nr = 0;
@@ -865,6 +856,15 @@ static void xdebug_execute_internal(zend_execute_data *current_execute_data, zva
 #else
 	void                (*tmp_error_cb)(int type, const char *error_filename, const uint32_t error_lineno, const char *format, va_list args) ZEND_ATTRIBUTE_PTR_FORMAT(printf, 4, 0) = NULL;
 #endif
+
+	if (XG(globals.library.mode) == XDEBUG_MODE_OFF) {
+		if (xdebug_old_execute_internal) {
+			xdebug_old_execute_internal(current_execute_data, return_value);
+		} else {
+			execute_internal(current_execute_data, return_value);
+		}
+		return;
+	}
 
 	XG_BASE(level)++;
 	if ((signed long) XG_BASE(level) > XINI_BASE(max_nesting_level) && (XINI_BASE(max_nesting_level) != -1)) {