--- a/xdebug.c
+++ b/xdebug.c
@@ -626,12 +626,12 @@ PHP_MINFO_FUNCTION(xdebug)
 
 ZEND_DLEXPORT void xdebug_statement_call(zend_execute_data *frame)
 {
+	zend_op_array *op_array = &frame->func->op_array;
+	int                   lineno;
+
 	if (EXPECTED(XG(globals.library.mode) == XDEBUG_MODE_OFF)) {
 		return;
 	}
-
-	zend_op_array *op_array = &frame->func->op_array;
-	int                   lineno;
 
 	if (!EG(current_execute_data)) {
 		return;