@derickr 2020-09-28
Thanks for looking into this. I'm closing this PR though as this is not the best place to do this. With xdebug.mode=off, Xdebug should not even override the zend_execute and friends calls in MINIT. Just overriding them switches PHP's engine to a less efficient execution model.

There is a ticket for me to track fixing this in the right way for Xdebug 3: https://bugs.xdebug.org/1813, which you are of course more than welcome to tackle. Feel free to reach out as well.