@derickr 2020-10-08
FWIW, I implemented this in a different way (PR #665), which should be even better performance wise. It'd be interesting to see what your benchmarks say on that, if you can still run them.