@igorwwwwwwwwwwwwwwwwwwww 2020-10-08
@derickr Very nice, here's the benchmark for current master:

```
$ hyperfine --warmup 1 \
    'php bench.php' \
    'php -d zend_extension=/Users/igor/code/xdebug/modules/xdebug.so -d xdebug.mode=off bench.php'

Benchmark #1: php bench.php
  Time (mean ± σ):      1.921 s ±  0.025 s    [User: 1.887 s, System: 0.019 s]
  Range (min … max):    1.904 s …  1.985 s    10 runs

Benchmark #2: php -d zend_extension=/Users/igor/code/xdebug/modules/xdebug.so -d xdebug.mode=off bench.php
  Time (mean ± σ):      1.935 s ±  0.030 s    [User: 1.898 s, System: 0.021 s]
  Range (min … max):    1.909 s …  2.010 s    10 runs

Summary
  'php bench.php' ran
    1.01 ± 0.02 times faster than 'php -d zend_extension=/Users/igor/code/xdebug/modules/xdebug.so -d xdebug.mode=off bench.php'
```

So the overhead is now negligible (possibly measurement noise).

Hopefully the results from this MR can still be helpful to potentially improve performance for some of the modes when xdebug _is_ enabled. Since not all code paths may be needed in all modes.

For example, `xdebug_add_stack_frame` may not be needed when running in remote debugging mode. Some of xdebug's functionality could be enabled on-demand on a per-request basis, etc.

Anyways, I'll leave it up to you what to do with the data. :)